Reorganize encryption code
